fixed users on render page
