+++ client/js/users and login/renderLoginSection.js
@@ -1,0 +1,28 @@
+import createElement from '../../lib/createElement.mjs';
+import loginUser from './loginUser.mjs';
+
+export default function renderLoginSection() {
+        const loginContainer = createElement('section', 'loginContainer', 'loginContainer')
+
+        chatSection.appendChild(loginContainer)
+
+        const loginPageHeader = createElement('h2', 'loginPageHeader', 'loginPageHeader', 'Select a Nickname')
+
+        
+        const userNameInput = createElement('input', 'userNameInput', 'userNameInput', '')
+        userNameInput.placeholder = 'Select a nickname';
+
+
+        const loginBtn = createElement('button', 'loginBtn', 'loginBtn', 'Login')
+
+        loginContainer.append(loginPageHeader, userNameInput, loginBtn)
+
+        loginBtn.addEventListener('click', () => {
+            console.log(userNameInput.value.trim());
+            loginUser(userNameInput.value.trim())
+            loginContainer.remove();
+        })
+
+    
+        
+}
