+++ server/userConnections/activeUsers.js
@@ -1,0 +1,6 @@
+module.exports = (io, socket, users, usersReady) => { 
+
+    socket.on('activeUsers', () => {
+        socket.emit('sendActiveUsers', users)
+    })
+}
